requirements.txt
# --- Core Frameworks ---
streamlit==1.39.0
uvicorn==0.23.2
python-multipart==0.0.20
python-dotenv==1.1.0
PyPDF2==3.0.1
typing-extensions==4.14.1
pydantic==2.11.7

# --- LangChain Core ---
langchain==0.3.26
langchain-core==0.3.74
langchain-community==0.3.26
langchain-text-splitters==0.3.9

# --- LangChain Integrations ---
langchain-huggingface==0.3.1
langchain-groq==0.3.7
langchain-pinecone==0.2.8

# --- Models & Vector Databases ---
huggingface-hub==0.34.4
transformers==4.55.2
pinecone==7.3.0
sentence-transformers==3.2.1
torch==2.5.1
grpcio==1.68.0
google-api-core==2.23.0
tqdm==4.67.1

requests==2.32.3
numpy==2.1.3
pandas==2.2.3

# --- Optional (Testing & Tools) ---
langchain-tests==0.3.20

# --- PostgreSQL & Neon Database ---
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
pgvector==0.3.6
asyncpg==0.30.0

# --- Web Search ---
tavily-python==0.5.0

